+++ NEWS.md
@@ -1,3 +1,5 @@
 # daedalus.workshop 0.0.1
 
 * This project now includes a [`NEWS.md`](https://r-pkgs.org/other-markdown.html#sec-news) file to inform users about changes and new features.
+
+* Initial package function `make_materials()` with templates for handout and phase 2 presentation, along with basic tests.